Workaround for issue 1 (https://code.google.com/p/hamcrest-reflection/issues/detail?id=1): "listsMatchingNestedClassesInReverseInheritanceOrder" failing on Java 7
Justification: If I understand correctly, the assumption that the Reflection API guarantees any particular order is simply false. In order to build the otherwise correct code and publish it on Maven Central, I decided to simply ignore this test for now and instead added the above issue to the tracker.